django/requirements.txt
django==4.1.3
djangorestframework==3.14.0
django-dry-rest-permissions==1.2.0
django-environ==0.9.0
psycopg2==2.9.5
django-cors-headers==3.13.0
